+++ pkg/services/accesscontrol/middleware.go
@@ -1,35 +1,20 @@
-package middleware
+package accesscontrol
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"strconv"
 	"time"
 
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/accesscontrol"
-	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/setting"
 	"github.com/grafana/grafana/pkg/util"
 	"github.com/grafana/grafana/pkg/web"
 )
 
-func authorize(c *models.ReqContext, ac accesscontrol.AccessControl, user *models.SignedInUser, evaluator accesscontrol.Evaluator) {
-	injected, err := evaluator.MutateScopes(c.Req.Context(), accesscontrol.ScopeInjector(buildScopeParams(c)))
-	if err != nil {
-		c.JsonApiErr(http.StatusInternalServerError, "Internal server error", err)
-		return
-	}
-
-	hasAccess, err := ac.Evaluate(c.Req.Context(), user, injected)
-	if !hasAccess || err != nil {
-		Deny(c, injected, err)
-		return
-	}
-}
-
-func Middleware(ac accesscontrol.AccessControl) func(web.Handler, accesscontrol.Evaluator) web.Handler {
-	return func(fallback web.Handler, evaluator accesscontrol.Evaluator) web.Handler {
+func Middleware(ac AccessControl) func(web.Handler, Evaluator) web.Handler {
+	return func(fallback web.Handler, evaluator Evaluator) web.Handler {
 		if ac.IsDisabled() {
 			return fallback
 		}
@@ -40,7 +25,24 @@ func Middleware(ac accesscontrol.AccessControl) func(web.Handler, accesscontrol.
 	}
 }
 
-func Deny(c *models.ReqContext, evaluator accesscontrol.Evaluator, err error) {
+func authorize(c *models.ReqContext, ac AccessControl, user *models.SignedInUser, evaluator Evaluator) {
+	injected, err := evaluator.MutateScopes(c.Req.Context(), ScopeInjector(ScopeParams{
+		OrgID:     c.OrgId,
+		URLParams: web.Params(c.Req),
+	}))
+	if err != nil {
+		c.JsonApiErr(http.StatusInternalServerError, "Internal server error", err)
+		return
+	}
+
+	hasAccess, err := ac.Evaluate(c.Req.Context(), user, injected)
+	if !hasAccess || err != nil {
+		deny(c, injected, err)
+		return
+	}
+}
+
+func deny(c *models.ReqContext, evaluator Evaluator, err error) {
 	id := newID()
 	if err != nil {
 		c.Logger.Error("Error from access control system", "error", err, "accessErrorID", id)
@@ -59,13 +61,18 @@ func Deny(c *models.ReqContext, evaluator accesscontrol.Evaluator, err error) {
 		return
 	}
 
+	message := ""
+	if evaluator != nil {
+		message = evaluator.String()
+	}
+
 	// If the user triggers an error in the access control system, we
 	// don't want the user to be aware of that, so the user gets the
 	// same information from the system regardless of if it's an
 	// internal server error or access denied.
 	c.JSON(http.StatusForbidden, map[string]string{
 		"title":         "Access denied", // the component needs to pick this up
-		"message":       fmt.Sprintf("You'll need additional permissions to perform this action. Permissions needed: %s", evaluator.String()),
+		"message":       fmt.Sprintf("You'll need additional permissions to perform this action. Permissions needed: %s", message),
 		"accessErrorId": id,
 	})
 }
@@ -82,17 +89,13 @@ func newID() string {
 	return "ACE" + id
 }
 
-func buildScopeParams(c *models.ReqContext) accesscontrol.ScopeParams {
-	return accesscontrol.ScopeParams{
-		OrgID:     c.OrgId,
-		URLParams: web.Params(c.Req),
-	}
+type OrgIDGetter func(c *models.ReqContext) (int64, error)
+type userCache interface {
+	GetSignedInUserWithCacheCtx(ctx context.Context, query *models.GetSignedInUserQuery) error
 }
 
-type OrgIDGetter func(c *models.ReqContext) (int64, error)
-
-func AuthorizeInOrgMiddleware(ac accesscontrol.AccessControl, db sqlstore.Store) func(web.Handler, OrgIDGetter, accesscontrol.Evaluator) web.Handler {
-	return func(fallback web.Handler, getTargetOrg OrgIDGetter, evaluator accesscontrol.Evaluator) web.Handler {
+func AuthorizeInOrgMiddleware(ac AccessControl, cache userCache) func(web.Handler, OrgIDGetter, Evaluator) web.Handler {
+	return func(fallback web.Handler, getTargetOrg OrgIDGetter, evaluator Evaluator) web.Handler {
 		if ac.IsDisabled() {
 			return fallback
 		}
@@ -102,17 +105,17 @@ func AuthorizeInOrgMiddleware(ac accesscontrol.AccessControl, db sqlstore.Store)
 			userCopy := *(c.SignedInUser)
 			orgID, err := getTargetOrg(c)
 			if err != nil {
-				Deny(c, nil, fmt.Errorf("failed to get target org: %w", err))
+				deny(c, nil, fmt.Errorf("failed to get target org: %w", err))
 				return
 			}
-			if orgID == accesscontrol.GlobalOrgID {
+			if orgID == GlobalOrgID {
 				userCopy.OrgId = orgID
 				userCopy.OrgName = ""
 				userCopy.OrgRole = ""
 			} else {
 				query := models.GetSignedInUserQuery{UserId: c.UserId, OrgId: orgID}
-				if err := db.GetSignedInUserWithCacheCtx(c.Req.Context(), &query); err != nil {
-					Deny(c, nil, fmt.Errorf("failed to authenticate user in target org: %w", err))
+				if err := cache.GetSignedInUserWithCacheCtx(c.Req.Context(), &query); err != nil {
+					deny(c, nil, fmt.Errorf("failed to authenticate user in target org: %w", err))
 					return
 				}
 				userCopy.OrgId = query.Result.OrgId
@@ -122,7 +125,7 @@ func AuthorizeInOrgMiddleware(ac accesscontrol.AccessControl, db sqlstore.Store)
 
 			authorize(c, ac, &userCopy, evaluator)
 
-			// Set the signed in user permissions in that org
+			// Set the sign-ed in user permissions in that org
 			c.SignedInUser.Permissions = userCopy.Permissions
 		}
 	}
@@ -140,27 +143,17 @@ func UseOrgFromContextParams(c *models.ReqContext) (int64, error) {
 }
 
 func UseGlobalOrg(c *models.ReqContext) (int64, error) {
-	return accesscontrol.GlobalOrgID, nil
+	return GlobalOrgID, nil
 }
 
-// Disable returns http 404 if shouldDisable is set to true
-func Disable(shouldDisable bool) web.Handler {
-	return func(c *models.ReqContext) {
-		if shouldDisable {
-			c.Resp.WriteHeader(http.StatusNotFound)
-			return
-		}
-	}
-}
-
-func LoadPermissionsMiddleware(ac accesscontrol.AccessControl) web.Handler {
+func LoadPermissionsMiddleware(ac AccessControl) web.Handler {
 	return func(c *models.ReqContext) {
 		if ac.IsDisabled() {
 			return
 		}
 
 		permissions, err := ac.GetUserPermissions(c.Req.Context(), c.SignedInUser,
-			accesscontrol.Options{ReloadCache: false})
+			Options{ReloadCache: false})
 		if err != nil {
 			c.JsonApiErr(http.StatusForbidden, "", err)
 			return
@@ -169,6 +162,6 @@ func LoadPermissionsMiddleware(ac accesscontrol.AccessControl) web.Handler {
 		if c.SignedInUser.Permissions == nil {
 			c.SignedInUser.Permissions = make(map[int64]map[string][]string)
 		}
-		c.SignedInUser.Permissions[c.OrgId] = accesscontrol.GroupScopesByAction(permissions)
+		c.SignedInUser.Permissions[c.OrgId] = GroupScopesByAction(permissions)
 	}
 }