+++ public/app/features/explore/RichHistory/RichHistory.tsx
@@ -1,11 +1,10 @@
+import { debounce } from 'lodash';
 import React, { PureComponent } from 'react';
 
 import { SelectableValue } from '@grafana/data';
 import { Themeable, withTheme, TabbedContainer, TabConfig } from '@grafana/ui';
-import { SortOrder } from 'app/core/utils/richHistory';
+import { SortOrder, RichHistorySearchFilters, RichHistorySettings } from 'app/core/utils/richHistory';
 import { RichHistoryQuery, ExploreId } from 'app/types/explore';
-
-import { RichHistorySearchFilters, RichHistorySettings } from '../../../core/utils/richHistoryTypes';
 
 import { RichHistoryQueriesTab } from './RichHistoryQueriesTab';
 import { RichHistorySettingsTab } from './RichHistorySettingsTab';
@@ -27,9 +26,11 @@ export const sortOrderOptions = [
 export interface RichHistoryProps extends Themeable {
   richHistory: RichHistoryQuery[];
   richHistorySettings: RichHistorySettings;
-  richHistorySearchFilters: RichHistorySearchFilters;
+  richHistorySearchFilters?: RichHistorySearchFilters;
   updateHistorySettings: (settings: RichHistorySettings) => void;
   updateHistorySearchFilters: (exploreId: ExploreId, filters: RichHistorySearchFilters) => void;
+  loadRichHistory: (exploreId: ExploreId) => void;
+  clearRichHistoryResults: (exploreId: ExploreId) => void;
   deleteRichHistory: () => void;
   activeDatasourceInstance?: string;
   firstTab: Tabs;
@@ -43,12 +44,22 @@ class UnThemedRichHistory extends PureComponent<RichHistoryProps> {
     this.props.updateHistorySettings({ ...this.props.richHistorySettings, ...settingsToUpdate });
   };
 
-  updateFilters = (filtersToUpdate: Partial<RichHistorySearchFilters>) => {
-    this.props.updateHistorySearchFilters(this.props.exploreId, {
-      ...this.props.richHistorySearchFilters,
+  updateFilters = (filtersToUpdate?: Partial<RichHistorySearchFilters>) => {
+    const filters = {
+      ...this.props.richHistorySearchFilters!,
       ...filtersToUpdate,
-    });
+    };
+    this.props.updateHistorySearchFilters(this.props.exploreId, filters);
+    this.loadRichHistory();
   };
+
+  clearResults = () => {
+    this.props.clearRichHistoryResults(this.props.exploreId);
+  };
+
+  loadRichHistory = debounce(() => {
+    this.props.loadRichHistory(this.props.exploreId);
+  }, 300);
 
   onChangeRetentionPeriod = (retentionPeriod: SelectableValue<number>) => {
     if (retentionPeriod.value !== undefined) {
@@ -62,39 +73,9 @@ class UnThemedRichHistory extends PureComponent<RichHistoryProps> {
   toggleActiveDatasourceOnly = () =>
     this.updateSettings({ activeDatasourceOnly: !this.props.richHistorySettings.activeDatasourceOnly });
 
-  onSelectDatasourceFilters = (datasourceFilters: SelectableValue[]) => this.updateFilters({ datasourceFilters });
-
-  onChangeSortOrder = (sortOrder: SortOrder) => this.updateFilters({ sortOrder });
-
-  /* If user selects activeDatasourceOnly === true, set datasource filter to currently active datasource.
-   * Filtering based on datasource won't be available. Otherwise set to null, as filtering will be
-   * available for user.
-   */
-  initFilters() {
-    if (this.props.richHistorySettings.activeDatasourceOnly && this.props.activeDatasourceInstance) {
-      this.onSelectDatasourceFilters([
-        { label: this.props.activeDatasourceInstance, value: this.props.activeDatasourceInstance },
-      ]);
-    }
-  }
-
-  componentDidMount() {
-    this.initFilters();
-  }
-
-  /**
-   * Updating filters on didMount and didUpdate because we don't know when activeDatasourceInstance is ready
-   */
-  componentDidUpdate(prevProps: RichHistoryProps) {
-    if (this.props.activeDatasourceInstance !== prevProps.activeDatasourceInstance) {
-      this.initFilters();
-    }
-  }
-
   render() {
-    const { activeDatasourceOnly, retentionPeriod } = this.props.richHistorySettings;
-    const { datasourceFilters, sortOrder } = this.props.richHistorySearchFilters;
-    const { richHistory, height, exploreId, deleteRichHistory, onClose, firstTab } = this.props;
+    const { richHistory, height, exploreId, deleteRichHistory, onClose, firstTab, activeDatasourceInstance } =
+      this.props;
 
     const QueriesTab: TabConfig = {
       label: 'Query history',
@@ -102,12 +83,11 @@ class UnThemedRichHistory extends PureComponent<RichHistoryProps> {
       content: (
         <RichHistoryQueriesTab
           queries={richHistory}
-          sortOrder={sortOrder}
-          datasourceFilters={datasourceFilters}
-          activeDatasourceOnly={activeDatasourceOnly}
-          retentionPeriod={retentionPeriod}
-          onChangeSortOrder={this.onChangeSortOrder}
-          onSelectDatasourceFilters={this.onSelectDatasourceFilters}
+          updateFilters={this.updateFilters}
+          clearRichHistoryResults={() => this.props.clearRichHistoryResults(this.props.exploreId)}
+          activeDatasourceInstance={activeDatasourceInstance}
+          richHistorySettings={this.props.richHistorySettings}
+          richHistorySearchFilters={this.props.richHistorySearchFilters}
           exploreId={exploreId}
           height={height}
         />
@@ -121,11 +101,11 @@ class UnThemedRichHistory extends PureComponent<RichHistoryProps> {
       content: (
         <RichHistoryStarredTab
           queries={richHistory}
-          sortOrder={sortOrder}
-          datasourceFilters={datasourceFilters}
-          activeDatasourceOnly={activeDatasourceOnly}
-          onChangeSortOrder={this.onChangeSortOrder}
-          onSelectDatasourceFilters={this.onSelectDatasourceFilters}
+          activeDatasourceInstance={activeDatasourceInstance}
+          updateFilters={this.updateFilters}
+          clearRichHistoryResults={() => this.props.clearRichHistoryResults(this.props.exploreId)}
+          richHistorySettings={this.props.richHistorySettings}
+          richHistorySearchFilters={this.props.richHistorySearchFilters}
           exploreId={exploreId}
         />
       ),